exec chmod is bad juju
Just let Puppet do its job. That's what it's designed for. The file resource doesn't specify content, only owner and mode. So that's all that Puppet will correct when it's evaluated.

This is untested, so you may have to tweak it slightly, but this practice is much better than chaining a chmod exec like you're writing a shell script.